Fix three real bugs found by running training through the actual production path
Every prior fix in this investigation was only exercised via the pilot's
splice into a frozen pretrained model, never through the project's real
training entrypoint (scripts/train.py -> Trainer -> DiffusionMoETransformer).
Running a short but genuinely real training step through that exact path
(small model, real streamed data, real bf16 mixed precision -- the
project's own default) surfaced three more bugs the pilot's setup had
incidentally masked:

- models/moe_layer.py, routing/centroids.py: NumPy has no bfloat16 dtype,
  so unguarded .numpy() calls on bf16 tensors crashed outright under real
  mixed-precision training. Fixed with a .float() cast before conversion.
- routing/separation.py: torch.cdist has no bfloat16 implementation either,
  hit in both landmark_scale and centroid_separation_loss. Same fix.
- training/trainer.py: ExpertCentroids.clip_norm_() (the fix for the
  unbounded-centroid-growth bug found earlier in this investigation) was
  never actually called anywhere except the pilot script's own bespoke
  training loop -- Trainer.train_step() had no equivalent call, so real
  training through the production path had no protection against the exact
  failure mode that fix exists for. Now called automatically after every
  optimizer step, for every DiffusionMoELayer block present.

None of PilotMoEBlock's own splice hit any of these, since its
DtypeCastWrapper already forces fp32 at the block boundary for an unrelated
reason -- confirming training-based validation on one code path doesn't
guarantee another actually works. All four fixes confirmed with a clean,
complete 8-step real run through scripts/train.py end to end.

Also anchors a new /outputs/ gitignore entry (Hydra's per-run output
directory) the same way /data/ already is.

+++ src/diffusion_moe/models/moe_layer.py
@@ -85,7 +85,14 @@ class DiffusionMoELayer(nn.Module):
         model's device/dtype.
         """
         batch, seq_len, d_model = z.shape
-        z_flat = z.detach().reshape(-1, d_model).cpu().numpy()
+        # .float() before .numpy(): NumPy has no bfloat16 dtype, so this
+        # crashes outright under real bf16 mixed-precision training (the
+        # project's own default precision) without it -- found by actually
+        # running a real step through Trainer, not caught by any prior
+        # training-free diagnostic or by the pilot (whose separate
+        # DtypeCastWrapper/PilotMoEBlock already cast to fp32 for other
+        # reasons, incidentally avoiding this exact crash).
+        z_flat = z.detach().reshape(-1, d_model).float().cpu().numpy()
 
         should_refresh = int(self._step.item()) % self.centroid_refresh_steps == 0
         if should_refresh:
